+++ src/Symfony/Component/Notifier/Bridge/LineBot/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add the `ssl` DSN option to send requests over plain HTTP
+ * Add the `LineBotOptions` class to set the recipient per message
 
 7.2
 ---